changes_frame_by_frame; para-sort-sub fix ID overflow; sort_mario

+++ ftc_changes_frame_by_frame_yulun.py
@@ -5,8 +5,8 @@ from math import floor
 from tqdm import tqdm
 import os
 
-exp_id="sortma2mh0it007"
-num_frames=11
+exp_id=""
+num_frames=51 
 
 if num_frames%2!=1:
     print("num_frames should be an odd number")
@@ -17,25 +17,29 @@ frames_range=[i-num_frames//2 for i in range(num_frames)]
 # num_frames: 5 => [-2,-1,0,1,2]
 # and so on
 
+
+
+
 # video = "TrackEvalYulun/data/trackers/mot_challenge/FISHow_dp-train/Trainow_dp/outpy10000.mp4"
 # video = "TrackEvalYulun/data/trackers/mot_challenge/FISHow_dp-train/Trainow_dp/outpy10000_yolo.mp4"
-# video = "OfficalYolo_euclidean_rematch_match_thr_1/outpy_OfficalYolo_euclidean_rematch_match_thr_1_10000_yolo.mp4"
+
 
 # video = "/work/marioeduardo-a/ftc/FTC-2024-data/Train/train.mp4"
 # label_file = "/work/marioeduardo-a/ftc/FTC-2024-data/Train/train_gt_mot.txt"
 
-TrackerName = f"Train{exp_id}"
-trackers_folder = f"TrackEvalYulun/data/trackers/mot_challenge/FISH{exp_id}-train"
-tracker_folder = os.path.join(trackers_folder, TrackerName)
+# TrackerName = f"Train{exp_id}"
+# trackers_folder = f"TrackEvalYulun/data/trackers/mot_challenge/FISH{exp_id}-train"
+# tracker_folder = os.path.join(trackers_folder, TrackerName)
+tracker_folder = 'OfficalYolo_euclidean_rematch_match_thr_1'
 result_folder = os.path.join(tracker_folder, "data")
 tracker_config_file = os.path.join(tracker_folder, "custom-tracker.yaml")
 raw_result_file = os.path.join(result_folder, "raw.txt")
 result_file = os.path.join(result_folder, f"FISH{exp_id}.txt")
 match_file = os.path.join(tracker_folder, f"FISH{exp_id}-pedestrian-bestmatch.txt")
 changes_file = os.path.join(tracker_folder, f"FISH{exp_id}-pedestrian-changes.txt")
-changes_folder = os.path.join(tracker_folder, f"{exp_id}_changes{len(frames_range)}fr")
+changes_folder = os.path.join(tracker_folder, f"{tracker_folder}_changes{len(frames_range)}fr")
 # changes_folder= "TrackEvalYulun/data/trackers/mot_challenge/FISHow_dp-train/Trainow_dp/changes"
-video=os.path.join(tracker_folder, f"outpy_{exp_id}_10000_yolo.mp4")
+video = f"{tracker_folder}/outpy_OfficalYolo_euclidean_rematch_match_thr_1_10000_yolo.mp4"
 
 # range_arr=[0 for i in range(frames_range)]
 if not os.path.exists(changes_folder):
@@ -61,8 +65,10 @@ with open(changes_file, 'r') as changes:
                 cap.set(cv2.CAP_PROP_POS_FRAMES, max(0,frame+j-1))
                 ret, img = cap.read()
                 tag="c" if j==0 else ""
-                cv2.imwrite(os.path.join(changes_folder,f"{frame+j:05d}{tag}.jpeg"), img)
+
+                # check if the same frame with a tag does not exist
+                if not os.path.exists(os.path.join(changes_folder,f"{frame+j:05d}c.jpeg")):
+                    cv2.imwrite(os.path.join(changes_folder,f"{frame+j:05d}{tag}.jpeg"), img)
         # print('i',i)
-     
 
 cap.release()
